Move ABC to util, made helper for exception printing
